refactor(scraping): Extract feed workflow
Move the home-feed workflow and its bounded listener drain to
`feed.FeedScraper`, consuming the browser-free payload policy that already
lives in `feed_payload`. The relocation is mechanical: listener registration
before navigation, removal of the identical callable in `finally`, drainage
even when that removal raises, the two-second settle, cancellation of every
unfinished read on every exit from it, the shielded one-second wait, the
leftover report and the final unshielded checkpoint are unchanged, as are the
scroll, stale and reference caps. `extract_feed` stays on the facade as a
delegate, and the shared rate-limit, modal and sleep boundaries now go through
`ScrapingSession`.

The drain, listener-lifecycle and tool deadline-consumer tests move into
`tests/scraping/test_feed.py` against the canonical owner, with two new cases:
the removed listener is the registered object, and the empty fast path never
suspends.

See also: #853
