+++ scripts/check_scraping_migration_manifest.py
@@ -283,7 +283,6 @@ _EXPLICIT_CALLER_CONTEXTS: dict[tuple[str, str, str], tuple[str, ...]] = {
         "boundaries",
         "detect_rate_limit",
     ): (
-        "_extract_feed_body",
         "get_sidebar_profiles",
         "_extract_search_page_once",
         "_extract_saved_jobs_page_once",
@@ -299,7 +298,6 @@ _EXPLICIT_CALLER_CONTEXTS: dict[tuple[str, str, str], tuple[str, ...]] = {
         "boundaries",
         "handle_modal_close",
     ): (
-        "_extract_feed_body",
         "get_sidebar_profiles",
         "_extract_search_page_once",
         "_extract_saved_jobs_page_once",
@@ -325,7 +323,6 @@ _EXPLICIT_CALLER_CONTEXTS: dict[tuple[str, str, str], tuple[str, ...]] = {
         "boundaries",
         "build_issue_diagnostics",
     ): (
-        "extract_feed",
         "scrape_person",
         "scrape_company",
         "_extract_search_page",
